% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_heatmap_section}
\alias{add_heatmap_section}
\title{Add a heatmap section to the report}
\usage{
add_heatmap_section(doc, wide_data, taxa_lookup, title, fig_num, cleanup)
}
\description{
Add a heatmap section to the report
}
\keyword{internal}
