% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_plot_to_doc}
\alias{add_plot_to_doc}
\title{Save a ggplot to a temp file and add to document}
\usage{
add_plot_to_doc(
  doc,
  plot,
  temp_files,
  width,
  height,
  display_width,
  display_height
)
}
\description{
Save a ggplot to a temp file and add to document
}
\keyword{internal}
