% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_station_sections}
\alias{add_station_sections}
\title{Add station report sections to document}
\usage{
add_station_sections(doc, station_summary)
}
\description{
Add station report sections to document
}
\keyword{internal}
