% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algaware-package.R
\docType{package}
\name{algaware-package}
\alias{algaware}
\alias{algaware-package}
\title{algaware: Interactive IFCB Phytoplankton Reporting Tool}
\description{
A Shiny application for processing, validating, and reporting phytoplankton data from Imaging FlowCytobot (IFCB) instruments. Integrates with IFCB Dashboard for data retrieval, supports interactive validation of classifier predictions, and generates Word reports with biomass plots, heatmaps, and image mosaics.
}
\author{
\strong{Maintainer}: Anders Torstensson \email{anders.torstensson@smhi.se}

}
\keyword{internal}
