% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{collect_ferrybox_data}
\alias{collect_ferrybox_data}
\title{Collect ferrybox data for station visits}
\usage{
collect_ferrybox_data(sample_times, ferrybox_path)
}
\arguments{
\item{sample_times}{POSIXct vector of sample timestamps.}

\item{ferrybox_path}{Path to ferrybox data folder.}
}
\value{
A data.frame with chlorophyll and other measurements per timestamp.
}
\description{
Collect ferrybox data for station visits
}
