% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaics.R
\name{compute_median_color}
\alias{compute_median_color}
\title{Compute median pixel color across images}
\usage{
compute_median_color(img_list)
}
\arguments{
\item{img_list}{List of magick image objects.}
}
\value{
Hex color string.
}
\description{
Compute median pixel color across images
}
\keyword{internal}
