% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{compute_sample_volumes}
\alias{compute_sample_volumes}
\title{Compute sample volumes per station visit}
\usage{
compute_sample_volumes(all_data)
}
\arguments{
\item{all_data}{Data.frame with sample, visit_id, STATION_NAME,
ml_analyzed, sample_time columns.}
}
\value{
Data.frame with visit_id, STATION_NAME, total_ml_analyzed,
median_time columns.
}
\description{
Compute sample volumes per station visit
}
\keyword{internal}
