% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_biomass_maps}
\alias{create_biomass_maps}
\title{Create biomass and chlorophyll maps}
\usage{
create_biomass_maps(station_summary)
}
\arguments{
\item{station_summary}{Aggregated station data from
\code{aggregate_station_data()}.}
}
\value{
A list with \code{biomass_map} and \code{chl_map} ggplot objects.
}
\description{
Create biomass and chlorophyll maps
}
