% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_settings}
\alias{default_settings}
\title{Default settings for algaware}
\usage{
default_settings()
}
\value{
A named list of default settings
}
\description{
Default settings for algaware
}
\keyword{internal}
