% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_config_dir}
\alias{get_config_dir}
\title{Get the configuration directory for algaware}
\usage{
get_config_dir()
}
\value{
Path to the configuration directory
}
\description{
Get the configuration directory for algaware
}
\keyword{internal}
