% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_settings_path}
\alias{get_settings_path}
\title{Get path to settings file}
\usage{
get_settings_path()
}
\value{
Path to the settings JSON file
}
\description{
Get path to settings file
}
\keyword{internal}
