% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaics.R
\name{get_taxon_rois}
\alias{get_taxon_rois}
\title{Get ROI information for a specific taxon in a set of samples}
\usage{
get_taxon_rois(classifications, taxa_lookup, taxon_name, sample_ids)
}
\arguments{
\item{classifications}{Classification data.frame with \code{sample_name},
\code{roi_number}, \code{class_name}.}

\item{taxa_lookup}{Taxa lookup table.}

\item{taxon_name}{Scientific name of the taxon.}

\item{sample_ids}{Character vector of sample PIDs to search within.}
}
\value{
A data.frame subset of classifications matching the taxon.
}
\description{
Get ROI information for a specific taxon in a set of samples
}
