% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{identify_diatom_classes}
\alias{identify_diatom_classes}
\title{Identify diatom classes from taxa lookup}
\usage{
identify_diatom_classes(taxa_lookup)
}
\arguments{
\item{taxa_lookup}{A data.frame with \code{clean_names} column.}
}
\value{
Character vector of class names likely to be diatoms.
}
\description{
Identify diatom classes from taxa lookup
}
\keyword{internal}
