% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{init_db_schema}
\alias{init_db_schema}
\title{Initialize the annotations database schema}
\usage{
init_db_schema(con)
}
\arguments{
\item{con}{A DBI connection object.}
}
\value{
Invisible NULL.
}
\description{
Creates tables compatible with ClassiPyR's schema.
}
\keyword{internal}
