% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaics.R
\name{justify_row}
\alias{justify_row}
\title{Justify a row of images to a target width}
\usage{
justify_row(imgs_row, widths_row, target_row_width, target_height, bg_color)
}
\arguments{
\item{imgs_row}{List of magick image objects.}

\item{widths_row}{Numeric vector of image widths.}

\item{target_row_width}{Target total row width in pixels.}

\item{target_height}{Row height in pixels.}

\item{bg_color}{Background fill color (hex string).}
}
\value{
A single magick image for the row.
}
\description{
Justify a row of images to a target width
}
\keyword{internal}
