% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{merge_ferrybox_data}
\alias{merge_ferrybox_data}
\title{Collect and merge ferrybox chlorophyll data}
\usage{
merge_ferrybox_data(config, matched, station_summary)
}
\description{
Collect and merge ferrybox chlorophyll data
}
\keyword{internal}
