% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{mod_data_loader_server}
\alias{mod_data_loader_server}
\title{Data Loader Module Server}
\usage{
mod_data_loader_server(id, config, rv)
}
\arguments{
\item{id}{Module namespace ID.}

\item{config}{Reactive values with settings.}

\item{rv}{Reactive values for app state.}
}
\value{
NULL (side effects only).
}
\description{
Data Loader Module Server
}
