% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{mod_data_loader_ui}
\alias{mod_data_loader_ui}
\title{Data Loader Module UI}
\usage{
mod_data_loader_ui(id)
}
\arguments{
\item{id}{Module namespace ID. Shiny modules use namespaced IDs so that
multiple instances of the same module don't conflict. \code{NS(id)}
creates a function that prefixes all input/output IDs with this namespace.}
}
\value{
A UI element.
}
\description{
Sidebar controls for selecting a cruise or date range and loading data.
}
