% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_frontpage.R
\name{mod_frontpage_server}
\alias{mod_frontpage_server}
\title{Front Page Mosaic Module Server}
\usage{
mod_frontpage_server(id, rv, config)
}
\arguments{
\item{id}{Module namespace ID.}

\item{rv}{Reactive values for app state.}

\item{config}{Reactive values with settings.}
}
\value{
NULL (side effects only).
}
\description{
Front Page Mosaic Module Server
}
