% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gallery.R
\name{mod_gallery_server}
\alias{mod_gallery_server}
\title{Gallery Module Server}
\usage{
mod_gallery_server(id, rv, config)
}
\arguments{
\item{id}{Module namespace ID.}

\item{rv}{Reactive values for app state.}

\item{config}{Reactive values with settings.}
}
\value{
NULL (side effects only).
}
\description{
Gallery Module Server
}
