% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Report Module Server}
\usage{
mod_report_server(id, rv, config)
}
\arguments{
\item{id}{Module namespace ID.}

\item{rv}{Reactive values for app state.}

\item{config}{Reactive values with settings.}
}
\value{
NULL (side effects only).
}
\description{
Report Module Server
}
