% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_samples.R
\name{mod_samples_ui}
\alias{mod_samples_ui}
\title{Sample Exclusion Module UI}
\usage{
mod_samples_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A UI element.
}
\description{
Sample Exclusion Module UI
}
