% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{mod_settings_server}
\alias{mod_settings_server}
\title{Settings Module Server}
\usage{
mod_settings_server(id, config)
}
\arguments{
\item{id}{Module namespace ID.}

\item{config}{Reactive values object for settings.}
}
\value{
NULL (side effects only).
}
\description{
Settings Module Server
}
