% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{mod_settings_ui}
\alias{mod_settings_ui}
\title{Settings Module UI}
\usage{
mod_settings_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A UI element.
}
\description{
Settings Module UI
}
