% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validation.R
\name{mod_validation_ui}
\alias{mod_validation_ui}
\title{Validation Module UI}
\usage{
mod_validation_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A UI element.
}
\description{
Validation Module UI
}
