% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gallery.R
\name{paginate_images}
\alias{paginate_images}
\title{Paginate a data frame of images}
\usage{
paginate_images(imgs, page, page_size)
}
\arguments{
\item{imgs}{Data frame of images.}

\item{page}{Integer page number (1-based).}

\item{page_size}{Number of images per page.}
}
\value{
Row-subset of \code{imgs} for the requested page.
}
\description{
Paginate a data frame of images
}
\keyword{internal}
