% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{sanitize_error_msg}
\alias{sanitize_error_msg}
\title{Sanitize error message for user display}
\usage{
sanitize_error_msg(msg)
}
\description{
Sanitize error message for user display
}
\keyword{internal}
