% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_settings}
\alias{save_settings}
\title{Save settings to disk}
\usage{
save_settings(settings)
}
\arguments{
\item{settings}{A named list of settings to persist}
}
\value{
Invisible NULL
}
\description{
Save settings to disk
}
