% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings.R
\name{validate_settings_inputs}
\alias{validate_settings_inputs}
\title{Validate settings form inputs}
\usage{
validate_settings_inputs(url, ppm, n_taxa, n_imgs)
}
\arguments{
\item{url}{Dashboard URL string.}

\item{ppm}{Pixels-per-micron numeric value.}

\item{n_taxa}{Number of mosaic taxa.}

\item{n_imgs}{Number of mosaic images.}
}
\value{
Character error message or \code{NULL} if all valid.
}
\description{
Validate settings form inputs
}
\keyword{internal}
