% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{download_all_data}
\alias{download_all_data}
\title{Download raw data, features, and classification files}
\usage{
download_all_data(config, sample_ids, storage)
}
\description{
Download raw data, features, and classification files
}
\keyword{internal}
