% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{filter_and_match}
\alias{filter_and_match}
\title{Filter and match metadata to stations}
\usage{
filter_and_match(
  dashboard_metadata,
  selection_mode,
  cruise,
  date_range,
  extra_stations
)
}
\description{
Filter and match metadata to stations
}
\keyword{internal}
