% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{mod_data_loader_ui}
\alias{mod_data_loader_ui}
\title{Data Loader Module UI}
\usage{
mod_data_loader_ui(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\value{
A UI element.
}
\description{
Data Loader Module UI
}
