% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validation.R
\name{mod_validation_server}
\alias{mod_validation_server}
\title{Validation Module Server}
\usage{
mod_validation_server(id, rv, config)
}
\arguments{
\item{id}{Module namespace ID.}

\item{rv}{Reactive values for app state.}

\item{config}{Reactive values with settings.}
}
\value{
NULL (side effects only).
}
\description{
Validation Module Server
}
