% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_loader.R
\name{resolve_classes}
\alias{resolve_classes}
\title{Resolve the class list from database or auto-generate}
\usage{
resolve_classes(config, taxa_lookup, classifications)
}
\description{
Resolve the class list from database or auto-generate
}
\keyword{internal}
